- Middleware related has been moved to `FluentlyHttpClient.Middleware` namespace, this effects:   - `LoggerHttpMiddleware`   - `TimerHttpMiddleware`   - `IFluentHttpMiddleware`   - `IFluentHttpMiddlewareRunner`
