+++ src/FluentlyHttpClient/FluentHttpClient.cs
@@ -7,6 +7,7 @@ using System.Net.Http;
 using System.Net.Http.Formatting;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
+using FluentlyHttpClient.Middleware;
 
 namespace FluentlyHttpClient
 {
