+++ src/FluentlyHttpClient/Middleware/FluentHttpMiddlewareRunner.cs
@@ -1,15 +1,28 @@
-using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Microsoft.Extensions.DependencyInjection;
 
-namespace FluentlyHttpClient
+namespace FluentlyHttpClient.Middleware
 {
+	/// <summary>
+	/// Http middleware runner (executor).
+	/// </summary>
 	public interface IFluentHttpMiddlewareRunner
 	{
+		/// <summary>
+		/// Run specified middlewares, and finally send the given request.
+		/// </summary>
+		/// <param name="middleware">Middleware to pipe.</param>
+		/// <param name="request">Request to send.</param>
+		/// <param name="send">Actual send function.</param>
+		/// <returns>Returns response.</returns>
 		Task<FluentHttpResponse> Run(IList<Type> middleware, FluentHttpRequest request, FluentHttpRequestDelegate send);
 	}
 
+	/// <summary>
+	/// Http middleware runner default implementation.
+	/// </summary>
 	public class FluentHttpMiddlewareRunner : IFluentHttpMiddlewareRunner
 	{
 		private readonly IServiceProvider _serviceProvider;
